refactor: Renderer 분리 및 DetailModalButtonContainer로 onClick 이동
